debian/rules: Add code to build an AA64 enrollment ISO
We're not yet shipping any images that need it, but it is nice to be able
to produce such an ISO for manual testing purposes. At some point
we may want to start producing AAVMF_VARS.{ms,snakeoil}.fd templates, but I
think ovmf-vars-generator is still x86-only, so we can't automate that yet.