build-ovmf: Break-up overly long command
The build-ovmf recipe is basically one long command so that everything
uses the environment provided by sourcing edksetup.sh. But really only
the 'build' commands need to run in that environement - we don't need it
when just moving files around. This does mean we need to source
edksetup.sh multiple times, but saves on indention and having to
add '; \' at the end of each line.